feat(layout): allow visualizer strip at top or bottom
